Add support for fly-by-wire to execute a debugging restart.
When issued the device will restart and then halt, fast-flashing the Pico LED until a USB serial device connects. Goal is to let a user see the startup messages in real time without needing to cache anything in memory. This may not be a bright idea but it seems to work OK for the moment.

This still needs to be documented properly.